requirements.txt
torch==2.5.1+cu124
torchvision==0.20.1+cu124
torchaudio==2.5.1+cu124
diffusers==0.32.0
transformers
accelerate
protobuf
sentencepiece
huggingface_hub
bitsandbytes
python-dotenv
--index-url https://download.pytorch.org/whl/cu124
--extra-index-url https://pypi.org/simple